requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
asyncpg==0.30.0
bcrypt==4.3.0
python-dotenv==1.1.0
anthropic==0.57.1
PyJWT==2.10.1
Pillow==10.3.0
pandas==2.3.1
numpy==1.26.4
matplotlib==3.10.5
seaborn==0.13.2
cycler==0.12.1
